+++ tests/test_directives.py
@@ -106,6 +106,24 @@ def test_current_api():
         hug.test.get(api, 'v3/version_call_tester').data
 
 
+def test_named_directives():
+    '''Ensure that it's possible to attach directives to named parameters'''
+    @hug.get()
+    def test(time:hug.directives.Timer=3):
+        return time
+
+    assert isinstance(test(), hug.directives.Timer)
+
+
+def test_named_directives_by_name():
+    '''Ensure that it's possible to attach directives to named parameters using only the name of the directive'''
+    @hug.get()
+    def test(time:__hug__.directive('timer')=3):
+        return time
+
+    assert isinstance(test(), hug.directives.Timer)
+
+
 def test_per_api_directives():
     '''Test to ensure it's easy to define a directive within an API'''
     @hug.directive(apply_globally=False)